PD-AI-main/pyproject.toml
[project]
name = "pd-fastapi-starter"
version = "0.1.0"
description = "Minimal FastAPI skeleton"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.123.9",
  "uvicorn>=0.38.0",
  "apscheduler>=3.10.4",
  "pydantic>=2.12.5",
  "PyJWT>=2.9.0",
  "bcrypt>=4.2.0",
  "requests>=2.32.0",
  "python-dotenv>=1.2.1",
  "python-multipart>=0.0.20",
  "Pillow>=12.0.0",
  "PyMySQL>=1.1.0",
]
